moved report code to its own fodler
